Add explicit human-readable flag and scale engine output
